pyproject.toml
[project]
name = "mapconverter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numba>=0.62.1",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
]
